server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
PyJWT==2.10.1
pwdlib[argon2]==0.2.1
email-validator==2.2.0
pytest==8.4.1
httpx==0.28.1
websockets==12.0
